removed some warnings
some further warning cleanup


Disabled warning 4215

Yes this is not a good idea. I will rewrite all involved interfaces at a later time when I actually have time for it.
working the warnings


working the build settings. not done yet


improved build settings


working on warnings


improvements

working on warnings
switching node id to 64bit
further cleanup

+++ src/Igor/src/iDataUpdateQueue.cpp
@@ -14,7 +14,7 @@ using namespace IgorAux;
 namespace Igor
 {
 
-    void iDataUpdateQueue::addNode(uint32 nodeID)
+    void iDataUpdateQueue::addNode(uint64 nodeID)
     {
         _mutex.lock();
