+++ src/Igor/src/iNodeFactory.h
@@ -69,11 +69,11 @@ namespace Igor
         {
             /*! node A to do an action with
             */
-            uint32 _nodeA = iNode::INVALID_NODE_ID;
+            uint64 _nodeA = iNode::INVALID_NODE_ID;
 
             /*! node B to do an action with
             */
-            uint32 _nodeB = iNode::INVALID_NODE_ID;
+            uint64 _nodeB = iNode::INVALID_NODE_ID;
 
             /*! the action to do with node A, B or both
             */
@@ -85,19 +85,19 @@ namespace Igor
         \param id id of ndoe
         \returns pointer to node
         */
-        iNode* getNode(uint32 id);
+        iNode* getNode(uint64 id);
 
         /*! \returns list of all node IDs of a certain node type
 
         \param nodeType type of nodes
         */
-        vector<uint32> getNodes(iNodeType nodeType);
+        vector<uint64> getNodes(iNodeType nodeType);
 
         /*! \returns true if node ID exists
 
         \param id the nodes ID
         */
-        bool isNode(uint32 id);
+        bool isNode(uint64 id);
 
         /*! create copy of node
 
@@ -119,7 +119,7 @@ namespace Igor
 
         \param nodeID id of node (asynchronously)
         */
-        void destroyNodeAsync(uint32 nodeID);
+        void destroyNodeAsync(uint64 nodeID);
 
         /*! applys asynchrounous actions to nodes
 
@@ -169,7 +169,7 @@ namespace Igor
 
         /*! mapping ids to nodes
         */
-        map<uint32, iNode*> _nodes;
+        map<uint64, iNode*> _nodes;
 
         /*! mutex to protect node list
         */
@@ -195,7 +195,7 @@ namespace Igor
         \param node the node to copy
         \param recursiveDepth recursive depth
         */
-        iNode* createCopyInternal(iNode* node, map<uint32, uint32>& nodeIDMap, uint32 recursiveDepth);
+        iNode* createCopyInternal(iNode* node, map<uint64, uint64>& nodeIDMap, uint32 recursiveDepth);
 
         /*! destroys node and all its children
 
@@ -209,7 +209,7 @@ namespace Igor
 
         \param nodeID id of node
         */
-        void destroyNode(uint32 nodeID);
+        void destroyNode(uint64 nodeID);
 
         /*! called once per frame by application
         */
